server/requirements.txt
flask==1.1.2
flask-cors==3.0.10
numpy
Pillow
matplotlib
deepface
